requirements.txt
black==19.3b0
celery==4.3.0
django==2.2.4
djangorestframework==3.10.2
psycopg2-binary==2.8.3
pylint==2.3.1
requests==2.22.0
redis==3.3.7
slackclient==2.1.0
